refactor: extract technician count summary readonly routes
